requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
PyMySQL==1.1.0
cryptography==41.0.4
APScheduler==3.10.4
python-dotenv==1.0.0
Werkzeug==2.3.7
